samples: peripheral: radio_test: refactor for multi-SoC radio support
Extract TX power control into radio_power_set.c, compiled only
when RADIO.TXPOWER is available (NRF_RADIO_HAS_TXPOWER).

Update radio_test.c for newer SoC peripheral layouts:
- Use EGU10/TIMER10/PHYEND as the default path (nRF54L, nRF71),
  keeping EGU0/TIMER0/END for nRF52/nRF53.
- Select ramp-up based on register capabilities instead of SoC series.
- Route TX/RX packet buffers through PACKETPTR or VDMACONFIG.LISTPTR
  depending on NRF_RADIO_HAS_PACKETPTR.

Expose channel_to_frequency() and radio_test_sweep_processing() for use
by the power control module.
